requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
openai>=1.50.0
python-telegram-bot[job-queue]>=21.6
